Add conn.request_path to google_cloud_logger format
Adding the request_path means that this value can be indexed
and filtered upon. Although the request_url attribute includes
the request path, the hostname in the request_url may not be a static value which
can make filtering on the request_path portion slow. Including the
request path as it's own top level attribute resolves this problem.